pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cfpmonitor"
version = "2026.2.0"
description = "Conference Call-for-Papers Policy Tracker and Comparison Tool"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Zhongkui Ma", email = "zhongkuima0419@gmail.com"}
]
dependencies = [
    "requests>=2.32.0",
    "beautifulsoup4>=4.14.0",
    "pyyaml>=6.0.3",
    "jinja2>=3.1.6",
    "ddgs>=9.10.0",
    "json-repair>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.15.0",
    "pre-commit>=4.0.0",
    "pytest>=9.0.2",
]

[tool.setuptools]
packages = ["src", "src.crawler", "src.extractor", "src.site"]

[tool.setuptools.package-data]
"src.extractor" = ["*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--verbose"]
